feat: 書籍管理トップ画面のUI修正
